Move device-plugin/ and images/ under runner/
