+++ runner/device-plugin/cmd/k8s-device-plugin/main.go
@@ -10,7 +10,7 @@ import (
 
 	"k8s.io/klog/v2"
 
-	"github.com/riseproject-dev/riscv-runner/device-plugin/pkg/plugin"
+	"github.com/riseproject-dev/riscv-runner/runner/device-plugin/pkg/plugin"
 )
 
 func main() {
