+++ runner/device-plugin/pkg/soc/detect.go
