+++ runner/images/README.md
@@ -7,28 +7,27 @@ For the full image inventory (every preinstalled tool with its version), the bui
 ## Layout
 
 ```
-images/
-├── runner/
-│   ├── Dockerfile.ubuntu              Runner image (multi-stage, parameterised by OS_VERSION)
-│   └── riscv-runner-entrypoint.sh     PID-1 entrypoint, exec's run.sh --jitconfig "$RUNNER_JITCONFIG"
-└── versions-map.json                  Mapping from Dockerfile ARGs to upstream version sources
+runner/images/
+├── Dockerfile.ubuntu              Runner image (multi-stage, parameterised by OS_VERSION)
+└── riscv-runner-entrypoint.sh     PID-1 entrypoint, exec's run.sh --jitconfig "$RUNNER_JITCONFIG"
 ```
 
-Companion files outside `images/`:
+Companion files outside `runner/images/`:
 
-- [`../scripts/update-versions.py`](../scripts/update-versions.py) — refreshes `versions-map.json` and the matching `ARG …_VERSION=` lines from the latest `actions/runner-images` release.
-- [`../.github/workflows/deploy-runner.yml`](../.github/workflows/deploy-runner.yml) — build, staging deploy, prod deploy.
-- [`../.github/workflows/update-images-versions-map.yml`](../.github/workflows/update-images-versions-map.yml) — weekly version sync.
+- [`../../runner/images/versions-map.json`](versions-map.json) — mapping from Dockerfile ARGs to upstream version sources (maintained by `update-versions.py`).
+- [`../../scripts/update-versions.py`](../../scripts/update-versions.py) — refreshes `versions-map.json` and the matching `ARG …_VERSION=` lines from the latest `actions/runner-images` release.
+- [`../../.github/workflows/deploy-runner.yml`](../../.github/workflows/deploy-runner.yml) — build, staging deploy, prod deploy.
+- [`../../.github/workflows/update-images-versions-map.yml`](../../.github/workflows/update-images-versions-map.yml) — weekly version sync.
 
 ## Build locally
 
 ```sh
 docker buildx build \
   --platform linux/riscv64 \
-  --file runner/Dockerfile.ubuntu \
+  --file Dockerfile.ubuntu \
   --build-arg OS_VERSION=24.04 \
   --tag riscv-runner:ubuntu-24.04-local \
-  runner
+  .
 ```
 
 Best run on a RISC-V host so no emulation is involved. On x86_64, `binfmt_misc` with QEMU will let the build complete, slowly.
@@ -39,7 +38,7 @@ Best run on a RISC-V host so no emulation is involved. On x86_64, `binfmt_misc` 
 python3 ../scripts/update-versions.py
 ```
 
-Reads the latest `ubuntu24/*` release of `actions/runner-images`, walks `versions-map.json`, and rewrites the matching `ARG …_VERSION=` lines in `runner/Dockerfile.ubuntu`. SHA256/SHA512 hashes are not updated automatically and must be edited by hand before merging.
+Reads the latest `ubuntu24/*` release of `actions/runner-images`, walks `versions-map.json`, and rewrites the matching `ARG …_VERSION=` lines in `Dockerfile.ubuntu`. SHA256/SHA512 hashes are not updated automatically and must be edited by hand before merging.
 
 The weekly workflow runs the same script and opens a draft PR if anything changes.
 
@@ -52,7 +51,7 @@ Each entry maps a Dockerfile ARG name to a field in the upstream runner-images m
   "arg": "PYTHON312_VERSION",
   "json_tool": "Cached Tools/Python",
   "match_prefix": "3.12",
-  "dockerfile": "images/runner/Dockerfile.ubuntu"
+  "dockerfile": "runner/images/Dockerfile.ubuntu"
 }
 ```
