+++ scripts/scw.py
@@ -2213,8 +2213,8 @@ runcmd:
     kubectl apply -f /etc/kubernetes/clusterroles.yml
 
     # Apply device plugin DaemonSets
-    curl -fsSL https://raw.githubusercontent.com/riseproject-dev/riscv-runner/refs/heads/main/device-plugin/k8s-ds-device-plugin.yaml | env TAG="@@TAG@@" envsubst | kubectl apply -f -
-    curl -fsSL https://raw.githubusercontent.com/riseproject-dev/riscv-runner/refs/heads/main/device-plugin/k8s-ds-node-labeller.yaml | env TAG="@@TAG@@" envsubst | kubectl apply -f -
+    curl -fsSL https://raw.githubusercontent.com/riseproject-dev/riscv-runner/refs/heads/main/runner/device-plugin/k8s-ds-device-plugin.yaml | env TAG="@@TAG@@" envsubst | kubectl apply -f -
+    curl -fsSL https://raw.githubusercontent.com/riseproject-dev/riscv-runner/refs/heads/main/runner/device-plugin/k8s-ds-node-labeller.yaml | env TAG="@@TAG@@" envsubst | kubectl apply -f -
 """
