+++ website/CLAUDE.md
@@ -44,8 +44,8 @@ Automatic on push to `main` via `.github/workflows/deploy-website.yml`. No manua
 ## Related directories in the monorepo
 
 - [`control-plane/`](https://github.com/riseproject-dev/riscv-runner/tree/main/control-plane): GitHub App webhook handler and scheduler
-- [`device-plugin/`](https://github.com/riseproject-dev/riscv-runner/tree/main/device-plugin): Kubernetes device plugin and node labeller
-- [`images/`](https://github.com/riseproject-dev/riscv-runner/tree/main/images): Runner container images
+- [`runner/device-plugin/`](https://github.com/riseproject-dev/riscv-runner/tree/main/runner/device-plugin): Kubernetes device plugin and node labeller
+- [`runner/images/`](https://github.com/riseproject-dev/riscv-runner/tree/main/runner/images): Runner container images
 
 Architecture pages reference source files in these directories. A separate [`riscv-runner-sample`](https://github.com/riseproject-dev/riscv-runner-sample) repo demonstrates user-facing usage and is not part of this monorepo.
